refa: rename mongo transformer to builder, rename sql escape to escape + escapePrimitive
