+++ packages/postgres/src/builder.ts
@@ -152,7 +152,7 @@ export class PostgresBuilder extends Builder {
 
   protected createElementQuery(key: string, value: any) {
     if (this.isJsonQuery(key)) {
-      return this.jsonContains(key, this.quote(JSON.stringify(value)))
+      return this.jsonContains(key, this.escape(value, 'json'))
     } else {
       return `${key} && ARRAY['${value}']::TEXT[]`
     }
@@ -216,17 +216,17 @@ export class PostgresBuilder extends Builder {
     return `'${value}'`
   }
 
-  escape(value: any, field?: Field) {
+  protected escapePrimitive(value: any) {
     if (value instanceof Date) {
       value = formatTime(value)
     } else if (value instanceof RegExp) {
       value = value.source
     } else if (value instanceof Buffer) {
       return `'\\x${value.toString('hex')}'::bytea`
-    } else if (!field && !!value && typeof value === 'object') {
+    } else if (!!value && typeof value === 'object') {
       return `${this.quote(JSON.stringify(value))}::jsonb`
     }
-    return super.escape(value, field)
+    return super.escapePrimitive(value)
   }
 
   toUpdateExpr(item: any, key: string, field?: Field, upsert?: boolean) {
@@ -258,7 +258,7 @@ export class PostgresBuilder extends Builder {
     // json_set cannot create deeply nested property when non-exist
     // therefore we merge a layout to it
     if (Object.keys(jsonInit).length !== 0) {
-      value = `(${value} || jsonb ${this.quote(JSON.stringify(jsonInit))})`
+      value = `(${value} || jsonb ${this.escape(jsonInit, 'json')})`
     }
 
     for (const prop in item) {
